split query and chat engine, add example notebook, add pip extra
