+++ docs/.gitignore
@@ -1,2 +1,4 @@
 /.quarto/
 lib64
+integrations/data
+integrations/storage
