+++ guardrails/integrations/llama_index/__init__.py
@@ -1,5 +1,8 @@
-from guardrails.integrations.llama_index.guardrails_engine import (
-    GuardrailsEngine,
+from guardrails.integrations.llama_index.guardrails_query_engine import (
+    GuardrailsQueryEngine,
+)
+from guardrails.integrations.llama_index.guardrails_chat_engine import (
+    GuardrailsChatEngine,
 )
 
-__all__ = ["GuardrailsEngine"]
+__all__ = ["GuardrailsQueryEngine", "GuardrailsChatEngine"]
