+++ pyproject.toml
@@ -60,6 +60,7 @@ guardrails-api = ">=0.1.0a1,<0.2.0"
 mlflow = {version = "^2.0.1", optional = true}
 uvloop = {version = "^0.20.0", optional = true}
 semver = "^3.0.2"
+llama-index = {version = ">=0.11.0", optional = true}
 
 [tool.poetry.extras]
 sql = ["sqlvalidator", "sqlalchemy", "sqlglot"]
@@ -71,6 +72,7 @@ huggingface = ["transformers", "torch", "jsonformer"]
 api = ["guardrails-api"]
 databricks = ["mlflow"]
 uv = ["uvloop"]
+llama = ["llama-index"]
 
 
 [tool.poetry.group.dev.dependencies]
